ticket:2680 for claudiosanchez - **Further commit** Correct type for a handful of loop variables Custom merge into /LFRic/trunk: r30884 cf. r30842 of /LFRic/branches/dev/claudiosanchez/r30803_SPT@30884
git-svn-id: https://metomi/svn/lfric.xm/LFRic/trunk@30885 4bbb03d4-c1cc-11e1-9ca4-71d347b01581

+++ um_physics/source/kernel/stph/spt_levels_cap_kernel_mod.F90
@@ -74,7 +74,7 @@ contains
     real(kind=r_def), intent(inout), dimension(undf_wth) :: dX
 
     real(kind=r_def) :: lev_amp_factor
-    real(kind=i_def) :: k
+    integer(kind=i_def) :: k
 
     ! Set to zero below spt_level_bottom
     do k = 0,spt_level_bottom-1
